+++ src/database/migrations/1744158981751-AddRegistradoInUserTable.ts
@@ -3,16 +3,23 @@ import { MigrationInterface, QueryRunner, TableColumn } from "typeorm";
 export class AddRegistradoInUserTable1744158981751 implements MigrationInterface {
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.addColumn("usuario", new TableColumn({
-            name: "registrado",
-            type: "boolean",
-            isNullable: false,
-            default: true
-        }));
+        const usuarioTable = await queryRunner.getTable("usuario");
+        
+        if (!usuarioTable.columns.find(c => c.name === "registrado")) {
+            await queryRunner.addColumn("usuario", new TableColumn({
+                name: "registrado",
+                type: "boolean",
+                isNullable: false,
+                default: true
+            }));
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropColumn("usuario", "registrado");
+        const usuarioTable = await queryRunner.getTable("usuario");
+        
+        if (usuarioTable.columns.find(c => c.name === "registrado")) {
+            await queryRunner.dropColumn("usuario", "registrado");
+        }
     }
-
 }
